+++ src/build-config/default-build-config.js
@@ -1,5 +1,6 @@
 import {CX_22_0} from '../version';
 import {LANDINGPAGE} from '../design-type';
+import DistFolder from '../dist-folder';
 
 /** @typedef {import('./build-config-interface').default} BuildConfigInterface */
 
@@ -13,6 +14,10 @@ export default class DefaultBuildConfig {
 
   get copyAssetsFolderPath() {
     return 'assets';
+  }
+
+  get assetResourceRuleFilename() {
+    return `${DistFolder.STATIC}/[name]-[contenthash][ext]`;
   }
 
   get designType() {