+++ test/templates/landingpage/content-elements/content/column-1/index.js
@@ -1,5 +1,8 @@
 const {cx, Icon} = require('@bsi-cx/design-build');
 
+/**
+ * @type {ContentElement}
+ */
 const element = cx.contentElement;
 
 module.exports = element;